+++ app/interactors/create_dublicate_for_gm_user.rb
@@ -11,6 +11,7 @@ class CreateDublicateForGmUser
     return if notification.nil?
     user_ids = context.delivery.deliveriable.head_users.pluck(:id)
     user_ids.each do |user_id|
+      next if Delivery.where(deliveriable_id: user_id, deliveriable_type: 'User', notification: notification).exists?
       delivery_params = { 'delivery_type' => 2, 'deliveriable_id' => user_id, 'deliveriable_type' => 'User', 'notification' => notification }
       delivery_param_params = { 'params' => { 'channel_id' => '' } }
       CreateDeliveryWithParams.call(delivery_params: delivery_params, delivery_param_params: delivery_param_params)